Update orders.d.ts
As you are stating in your comment " * Either type or options but not both may be present".
Currently Paypal doesn't allow type and options to be used at the same time, but types it paypal-js are enforcing developers to use both.
I want to use paypal-js types over defining empty type declaration in my local types folder. so i'm suggesting this fix to paypal-js types.